Update and rename WeatherStations.jsx to FarmPlotsMap.jsx
